Module fvm_actor_utils::receiver

Structs

ReceiverHook

Implements a guarded call to a token receiver hook

UniversalReceiverParams

Parameters for universal receiver

Enums

ReceiverHookError

Traits

RecipientData

UniversalReceiver

Standard interface for an actor that wishes to receive FRC-0046 tokens or other assets

Type Aliases

ReceiverType

Type of asset received - could be tokens (FRC46 or other) or other assets